requirements.txt
absl-py==2.1.0 ; python_full_version >= "3.12.7" and python_full_version < "4.0.0"
astunparse==1.6.3 ; python_full_version >= "3.12.7" and python_full_version < "4.0.0"
attrs==24.2.0 ; python_full_version >= "3.12.7" and python_full_version < "4.0.0"
certifi==2024.8.30 ; python_full_version >= "3.12.7" and python_full_version < "4.0.0"
cffi==1.17.1 ; python_full_version >= "3.12.7" and python_full_version < "4.0.0"
cfgv==3.4.0 ; python_full_version >= "3.12.7" and python_full_version < "4.0.0"
charset-normalizer==3.4.0 ; python_full_version >= "3.12.7" and python_full_version < "4.0.0"
contourpy==1.3.0 ; python_full_version >= "3.12.7" and python_full_version < "4.0.0"
coverage==7.6.4 ; python_full_version >= "3.12.7" and python_full_version < "4.0.0"
cycler==0.12.1 ; python_full_version >= "3.12.7" and python_full_version < "4.0.0"
distlib==0.3.9 ; python_full_version >= "3.12.7" and python_full_version < "4.0.0"
filelock==3.16.1 ; python_full_version >= "3.12.7" and python_full_version < "4.0.0"
flake8==6.1.0 ; python_full_version >= "3.12.7" and python_full_version < "4.0.0"
flatbuffers==24.3.25 ; python_full_version >= "3.12.7" and python_full_version < "4.0.0"
fonttools==4.54.1 ; python_full_version >= "3.12.7" and python_full_version < "4.0.0"
gast==0.6.0 ; python_full_version >= "3.12.7" and python_full_version < "4.0.0"
google-pasta==0.2.0 ; python_full_version >= "3.12.7" and python_full_version < "4.0.0"
grpcio==1.67.1 ; python_full_version >= "3.12.7" and python_full_version < "4.0.0"
h5py==3.12.1 ; python_full_version >= "3.12.7" and python_full_version < "4.0.0"
identify==2.6.1 ; python_full_version >= "3.12.7" and python_full_version < "4.0.0"
idna==3.10 ; python_full_version >= "3.12.7" and python_full_version < "4.0.0"
isort==5.13.2 ; python_full_version >= "3.12.7" and python_full_version < "4.0.0"
jax==0.4.35 ; python_full_version >= "3.12.7" and python_full_version < "4.0.0"
jaxlib==0.4.35 ; python_full_version >= "3.12.7" and python_full_version < "4.0.0"
keras==3.6.0 ; python_full_version >= "3.12.7" and python_full_version < "4.0.0"
kiwisolver==1.4.7 ; python_full_version >= "3.12.7" and python_full_version < "4.0.0"
libclang==18.1.1 ; python_full_version >= "3.12.7" and python_full_version < "4.0.0"
markdown-it-py==3.0.0 ; python_full_version >= "3.12.7" and python_full_version < "4.0.0"
markdown==3.7 ; python_full_version >= "3.12.7" and python_full_version < "4.0.0"
markupsafe==3.0.2 ; python_full_version >= "3.12.7" and python_full_version < "4.0.0"
matplotlib==3.9.2 ; python_full_version >= "3.12.7" and python_full_version < "4.0.0"
mccabe==0.7.0 ; python_full_version >= "3.12.7" and python_full_version < "4.0.0"
mdurl==0.1.2 ; python_full_version >= "3.12.7" and python_full_version < "4.0.0"
mediapipe==0.10.18 ; python_full_version >= "3.12.7" and python_full_version < "4.0.0"
ml-dtypes==0.4.1 ; python_full_version >= "3.12.7" and python_full_version < "4.0.0"
namex==0.0.8 ; python_full_version >= "3.12.7" and python_full_version < "4.0.0"
nodeenv==1.9.1 ; python_full_version >= "3.12.7" and python_full_version < "4.0.0"
numpy==1.26.4 ; python_full_version >= "3.12.7" and python_full_version < "4.0.0"
opencv-contrib-python==4.10.0.84 ; python_full_version >= "3.12.7" and python_full_version < "4.0.0"
opencv-python==4.10.0.84 ; python_full_version >= "3.12.7" and python_full_version < "4.0.0"
opt-einsum==3.4.0 ; python_full_version >= "3.12.7" and python_full_version < "4.0.0"
optree==0.13.0 ; python_full_version >= "3.12.7" and python_full_version < "4.0.0"
packaging==24.1 ; python_full_version >= "3.12.7" and python_full_version < "4.0.0"
pillow==11.0.0 ; python_full_version >= "3.12.7" and python_full_version < "4.0.0"
platformdirs==4.3.6 ; python_full_version >= "3.12.7" and python_full_version < "4.0.0"
pre-commit==3.8.0 ; python_full_version >= "3.12.7" and python_full_version < "4.0.0"
protobuf==4.25.5 ; python_full_version >= "3.12.7" and python_full_version < "4.0.0"
pycodestyle==2.11.1 ; python_full_version >= "3.12.7" and python_full_version < "4.0.0"
pycparser==2.22 ; python_full_version >= "3.12.7" and python_full_version < "4.0.0"
pyflakes==3.1.0 ; python_full_version >= "3.12.7" and python_full_version < "4.0.0"
pygame==2.6.1 ; python_full_version >= "3.12.7" and python_full_version < "4.0.0"
pygments==2.18.0 ; python_full_version >= "3.12.7" and python_full_version < "4.0.0"
pyparsing==3.2.0 ; python_full_version >= "3.12.7" and python_full_version < "4.0.0"
python-dateutil==2.9.0.post0 ; python_full_version >= "3.12.7" and python_full_version < "4.0.0"
pyyaml==6.0.2 ; python_full_version >= "3.12.7" and python_full_version < "4.0.0"
requests==2.32.3 ; python_full_version >= "3.12.7" and python_full_version < "4.0.0"
rich==13.9.4 ; python_full_version >= "3.12.7" and python_full_version < "4.0.0"
scipy==1.14.1 ; python_full_version >= "3.12.7" and python_full_version < "4.0.0"
sentencepiece==0.2.0 ; python_full_version >= "3.12.7" and python_full_version < "4.0.0"
setuptools==75.3.0 ; python_full_version >= "3.12.7" and python_full_version < "4.0.0"
six==1.16.0 ; python_full_version >= "3.12.7" and python_full_version < "4.0.0"
sounddevice==0.5.1 ; python_full_version >= "3.12.7" and python_full_version < "4.0.0"
tensorboard-data-server==0.7.2 ; python_full_version >= "3.12.7" and python_full_version < "4.0.0"
tensorboard==2.18.0 ; python_full_version >= "3.12.7" and python_full_version < "4.0.0"
tensorflow==2.18.0 ; python_full_version >= "3.12.7" and python_full_version < "4.0.0"
termcolor==2.5.0 ; python_full_version >= "3.12.7" and python_full_version < "4.0.0"
typing-extensions==4.12.2 ; python_full_version >= "3.12.7" and python_full_version < "4.0.0"
urllib3==2.2.3 ; python_full_version >= "3.12.7" and python_full_version < "4.0.0"
virtualenv==20.27.1 ; python_full_version >= "3.12.7" and python_full_version < "4.0.0"
werkzeug==3.1.2 ; python_full_version >= "3.12.7" and python_full_version < "4.0.0"
wheel==0.44.0 ; python_full_version >= "3.12.7" and python_full_version < "4.0.0"
wrapt==1.16.0 ; python_full_version >= "3.12.7" and python_full_version < "4.0.0"
